Softer value optimization
A variable key containing the string "01" was optimized to integer 1,
either the calculated string "1 " was converted to 1, so an expression
like "key + ' ' + name" results in a string which was not the desired
output. This behavior is now changed, so when "01" or "01 " is explicitly
wanted to be calculated, the function int() or float() must be invoked.